move the ones I missed
